Migrate tests from api-extractor-06 into api-extractor-scenarios
