+++ build-tests/api-extractor-scenarios/etc/test-outputs/defaultExportOfEntryPoint1/api-extractor-scenarios.api.ts
@@ -1,0 +1,6 @@
+// @public (undocumented)
+class DefaultClass {
+}
+
+
+// (No @packageDocumentation comment for this package)
