+++ build-tests/api-extractor-scenarios/etc/test-outputs/defaultExportOfEntryPoint1/rollup.d.ts
@@ -1,0 +1,7 @@
+
+/** @public */
+declare class DefaultClass {
+}
+export default DefaultClass;
+
+export { }
