+++ build-tests/api-extractor-scenarios/etc/test-outputs/defaultExportOfEntryPoint2/api-extractor-scenarios.api.ts
@@ -1,0 +1,5 @@
+// @public (undocumented)
+declare const defaultFunctionStatement: () => void;
+
+
+// (No @packageDocumentation comment for this package)
