+++ build-tests/api-extractor-scenarios/etc/test-outputs/defaultExportOfEntryPoint2/rollup.d.ts
@@ -1,0 +1,6 @@
+
+/** @public */
+declare const defaultFunctionStatement: () => void;
+export default defaultFunctionStatement;
+
+export { }
