+++ build-tests/api-extractor-scenarios/etc/test-outputs/defaultExportOfEntryPoint3/api-extractor-scenarios.api.ts
@@ -1,0 +1,5 @@
+// @public (undocumented)
+function defaultFunctionDeclaration(): void;
+
+
+// (No @packageDocumentation comment for this package)
