+++ build-tests/api-extractor-scenarios/etc/test-outputs/defaultExportOfEntryPoint3/rollup.d.ts
@@ -1,0 +1,8 @@
+
+/**
+ * @public
+ */
+declare function defaultFunctionDeclaration(): void;
+export default defaultFunctionDeclaration;
+
+export { }
