+++ build-tests/api-extractor-scenarios/etc/test-outputs/defaultExportOfEntryPoint4/api-extractor-scenarios.api.ts
@@ -1,0 +1,5 @@
+// @public (undocumented)
+declare const _default: "literal";
+
+
+// (No @packageDocumentation comment for this package)
