+++ build-tests/api-extractor-scenarios/etc/test-outputs/defaultExportOfEntryPoint4/rollup.d.ts
@@ -1,0 +1,5 @@
+
+declare const _default: "literal";
+export default _default;
+
+export { }
