+++ build-tests/api-extractor-scenarios/src/defaultExportOfEntryPoint2/index.ts
@@ -1,0 +1,7 @@
+// Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
+// See LICENSE in the project root for license information.
+
+/** @public */
+const defaultFunctionStatement = () => { };
+
+export default defaultFunctionStatement;
