+++ build-tests/api-extractor-scenarios/src/defaultExportOfEntryPoint3/index.ts
@@ -1,0 +1,7 @@
+// Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
+// See LICENSE in the project root for license information.
+
+/**
+ * @public
+ */
+export default function defaultFunctionDeclaration() { };
