+++ build-tests/api-extractor-scenarios/src/defaultExportOfEntryPoint4/index.ts
@@ -1,0 +1,5 @@
+// Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
+// See LICENSE in the project root for license information.
+
+/** @public */
+export default "literal";
