--- build-tests/api-extractor-test-06/api-extractor-template.json
@@ -1,20 +1,0 @@
-{
-  "$schema": "https://developer.microsoft.com/json-schemas/api-extractor/api-extractor.schema.json",
-  "compiler": {
-    "configType": "tsconfig",
-    "rootFolder": "."
-  },
-  "project": {
-    "entryPointSourceFile": ""
-  },
-  "dtsRollup": {
-    "enabled": true,
-    "mainDtsRollupPath": ""
-  },
-  "apiReviewFile": {
-    "enabled": false
-  },
-  "apiJsonFile": {
-    "enabled": false
-  }
-}
