--- build-tests/api-extractor-test-06/src/FunctionDeclaration.ts
@@ -1,4 +1,0 @@
-/**
- * @public
- */
-export default function defaultFunctionDeclaration() { };
