--- build-tests/api-extractor-test-06/src/FunctionStatement.ts
@@ -1,6 +1,0 @@
-/**
- * @public
- */
-const defaultFunctionStatement = () => { };
-
-export default defaultFunctionStatement;
