--- build-tests/api-extractor-test-06/src/Literal.ts
@@ -1,4 +1,0 @@
-/**
- * @public
- */
-export default "literal";
