--- build-tests/api-extractor-test-06/tsconfig-caller.json
@@ -1,9 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "outDir": ".",
-    "sourceMap": false,
-    "declaration": false
-  },
-  "files": ["caller.ts"],
-}
