--- build-tests/api-extractor-test-06/tsconfig.json
@@ -1,27 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es6",
-    "forceConsistentCasingInFileNames": true,
-    "module": "commonjs",
-    "declaration": true,
-    "sourceMap": true,
-    "experimentalDecorators": true,
-    "strictNullChecks": true,
-    "types": [
-      "node",
-      "jest"
-    ],
-    "lib": [
-      "es5",
-      "scripthost",
-      "es2015.collection",
-      "es2015.promise",
-      "es2015.iterable",
-      "dom"
-    ],
-    "outDir": "lib"
-  },
-  "include": [
-    "src/**/*.ts",
-  ]
-}
